--- frontend/static/src/utils/checkAuthentication.js
@@ -1,10 +1,0 @@
-
-const checkAuthentication = () => {
-    if (localStorage.getItem('current-user')) {
-        return true
-    }
-    else {
-        return false
-    }
-};
-export default checkAuthentication;
